Pull out inproc sockets for zmq backend in favor of a simple Queue
The main motivation here is that it doesn't make sense to keep
supporting a complex worker networking pattern when we don't ever intend
to use it. At this point it's legacy bloat that's ripe for removal.

We can first pull out the backend sockets and stick a Queue in their
place. There's a lot more refactoring that can be done here, but this is
a good stopping point.